% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apply_sign_cutoffs.R
\name{Apply_sign_cutoffs}
\alias{Apply_sign_cutoffs}
\title{Select significant elements}
\usage{
Apply_sign_cutoffs(df, individualPvalueCutoff = 0.3,
  combinedPvalueCutoff = 0.05, combinedFDRCutoff = 0.1)
}
\arguments{
\item{df, }{individualPvalueCutoff, combinedPvalueCutoff, combinedFDRCutoff}
}
\value{
A matrix with statistics for signifcant elements
}
\description{
This function applies the following significance cutoffs
individual p-value < 0.3; combined p-value < 0.05; fdr < 0.1
}

