% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_combined.R
\name{Calc_combined}
\alias{Calc_combined}
\title{Calc combined p value across expts.}
\usage{
Calc_combined(s_df)
}
\arguments{
\item{data}{frame}
}
\value{
A matrix with combined pvalue and fdr
}
\description{
This function code for calc combined p value across datasets.
}

