% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_cor.R
\name{Calc_cor}
\alias{Calc_cor}
\title{Calculate correlation between vectors}
\usage{
Calc_cor(pair, df, samples)
}
\arguments{
\item{pair, }{df, samples}
}
\value{
A matrix with statistics for each element pair
}
\description{
This function calculate correlation between vectors
}

