% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_median_val.R
\name{Calc_median_val}
\alias{Calc_median_val}
\title{Calculate median coefficient or fold change}
\usage{
Calc_median_val(df, pattern)
}
\arguments{
\item{df, }{pattern}
}
\value{
A data frame with statistics for each element
}
\description{
This function calculate median coefficient or fold change for each element
}

