% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_consistency.R
\name{Check_consistency}
\alias{Check_consistency}
\title{Check consistent direction}
\usage{
Check_consistency(s_df, pattern, Threshold, numbDatasets = 2)
}
\arguments{
\item{df, }{patternToSearch, threshold}
}
\value{
A matrix with elements showing consistent direction across datasets
}
\description{
This function selects elements showing consistent direction across datasets
}

