% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compare_groups.R
\name{Compare_groups}
\alias{Compare_groups}
\title{Compare groups}
\usage{
Compare_groups(mapFile, geneFile, sampleIdColName, factorColName, groupA,
  groupB, geneSymbolColName)
}
\arguments{
\item{mapFile, }{geneFile, sampleIdColName, factorColName, groupA, groupB, geneSymbolColName}
}
\value{
A data frame with statistics for each element
}
\description{
This function compares the values in two groups for each element
}

