% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_in_group.R
\name{Correlation_in_group}
\alias{Correlation_in_group}
\title{Correlation_in_group}
\usage{
Correlation_in_group(mapFile, geneFile, sampleIdColName, factorColName, groupA,
  geneSymbolColName, selected_genes, usepairs = NA)
}
\arguments{
\item{mapFile, }{geneFile, sampleIdColName, factorColName, groupA, geneSymbolColName, selected genes}
}
\value{
A data frame with statistics for each element
}
\description{
This function calculates correlation in a group for each element pair
}

