% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff_abundance.R
\name{Diff_abundance}
\alias{Diff_abundance}
\title{Compare abundance between groups for an element}
\usage{
Diff_abundance(element, df, samplesA, samplesB)
}
\arguments{
\item{element, }{df, samplesA, samplesB}
}
\value{
A matrix with statistics for each element
}
\description{
This function compares the values in two groups for an element
}

