% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_shortest_paths.R
\name{Get_shortest_paths}
\alias{Get_shortest_paths}
\title{Get counts for all shortest paths between given nodes in a bipartite networks}
\usage{
Get_shortest_paths(pair, myNetwork, sumAllFractionsForAllNodes)
}
\arguments{
\item{pair, }{myNetwork, sumAllFractionsForAllNodes}
}
\value{
Updates the row in pair,sumAllFractionsForAllNodes
}
\description{
This function calculates shortest paths between given nodes in a bipartite networks
}

