% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_template_matrix.R
\name{Get_template_matrix}
\alias{Get_template_matrix}
\title{Get_template_matrix}
\usage{
Get_template_matrix(SeedNodes, allPairs)
}
\arguments{
\item{FromNodes, }{allPairs}
}
\value{
A matrix with rownames (source-target) and column names as nodes from source nodes set.
}
\description{
This function creates a matrix , with each entry recording for each node(in a column), 
if it appears(not 1 or 0, but a normalized number, since there can be more 
than one shortest path between a pair of nodes) in the shortest path between a 
pair of nodes(in a row), the rows contain all pairs of nodes between two groups.
Thus the sum of each column is the betweeness centrality for the node(in that column)
}

