% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_subnetworks.R
\name{Identify_subnetworks}
\alias{Identify_subnetworks}
\title{Identify top subnetwork}
\usage{
Identify_subnetworks(outNetwork)
}
\arguments{
\item{network}{in which to identify subnetworks}
}
\value{
A data frame (network) that is Puc compatiable
}
\description{
This function returns the induced subgraph of the nodes in the top cluster using mcode
}

