% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Puc_compatiable_network.R
\name{Puc_compatible_network}
\alias{Puc_compatible_network}
\title{Calculate network that satisfies the PUC criteria}
\usage{
Puc_compatible_network(pairs_df, genes_df)
}
\arguments{
\item{pairs_df, }{genes_df}
}
\value{
A data frame (network) that is Puc compatiable
}
\description{
This function keeps edges that satisfy expected correlation and fold change relationships
}

